// src/interface/weeklyreport/type.ts
export interface IUpdateAttendanceMoreParams {
  params: IUpdateAttendanceMoreContentParams;
}

export interface IUpdateAttendanceMoreContentParams {
  args: IUpdateAttendanceMoreContentArg[][];
}

export interface IUpdateAttendanceMoreContentArg {
  employee_code: string;
  date: string;
  shift_name: string;
}

export interface WeeklyReportState {
  id?: number;
  date?: string;
  employee_code?: string;
  total_work_time?: number;
  shift_start?: number;
  shift_end?: number;
  total_shift_work_time?: number;
  holiday_start?: number;
  holiday_end?: number;
  rest_start?: number;
  rest_end?: number;
  employee_name?: string;
  shift_name?: string;
  explanation?: boolean;
  is_explanation?: boolean;
  type_shifts?: number;
  on_time?: boolean;
  wrong_time?: boolean;
  rest_shifts?: boolean;
  empty_shifts?: boolean;
  extra_hour?: number;
  approval_hour?: number;
  company?: string;
  department?: string;
  job_title?: string;
  attendance_attempt_1?: string;
  attendance_attempt_2?: string;
  attendance_attempt_3?: string;
  attendance_attempt_4?: string;
  attendance_attempt_5?: string;
  attendance_attempt_6?: string;
  attendance_attempt_7?: string;
  attendance_attempt_8?: string;
  attendance_attempt_9?: string;
  attendance_attempt_10?: string;
  attendance_attempt_11?: string;
  attendance_attempt_12?: string;
  attendance_attempt_13?: string;
  attendance_attempt_14?: string;
  attendance_attempt_15?: string;
  last_attendance_attempt?: string;
  time_keeping_code?: string;
  resource_calendar: number;
  missing_checkin_break: boolean;
  split_shift: boolean;
}
// ...
export interface IUpdateSchedulingMoreResult {
  id: number;
  day_of_week: number;
  total_attendance: number;
  create_uid: number;
  write_uid: number;
  employee_name: string;
  employee_code: string;
  time_keeping_code: string;
  department: string;
  company: string;
  job_title: string;
  shift_name: string;
  explaination_type: string;
  explaination_content: string;
  holiday_name: string;
  date: string;
  join_date: any;
  probation_completion_wage: any;
  is_worked: boolean;
  check_is_enough_time: boolean;
  missing_checkin_break: boolean;
  missing_checkout: boolean;
  missing_checkin: boolean;
  split_shift: boolean;
  is_holiday: boolean;
  is_old: boolean;
  attendance_attempt_1: string;
  attendance_attempt_2: string;
  attendance_attempt_3: string;
  attendance_attempt_4: any;
  attendance_attempt_5: any;
  attendance_attempt_6: any;
  attendance_attempt_7: any;
  attendance_attempt_8: any;
  attendance_attempt_9: any;
  attendance_attempt_10: any;
  attendance_attempt_11: any;
  attendance_attempt_12: any;
  attendance_attempt_13: any;
  attendance_attempt_14: any;
  attendance_attempt_15: any;
  last_attendance_attempt: string;
  holiday_start: string;
  holiday_end: string;
  create_date: string;
  write_date: string;
  surplus_minute_work: number;
  probation_wage_rate: number;
  standard_working_day: any;
  shift_start: number;
  shift_end: number;
  rest_start: number;
  rest_end: number;
  total_shift_work_time: number;
  ot_normal: number;
  ot_holiday: number;
  night_hours_normal: number;
  night_hours_holiday: number;
  minute_worked_day_normal: number;
  minute_worked_day_holiday: number;
  attendance_late: number;
  leave_early: number;
  total_work_time: number;
  severance_day: any;
  night_shift: any;
}